renamed index.ts to module.ts moved guards/interceptors into their own folder
