+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { DynamicModule, Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { ScheduleModule } from '@nestjs/schedule'
 
-import { ApiModule } from '@src/module.api'
+import { ApiModule } from '@src/module.api/module'
 import { DatabaseModule } from '@src/module.database/module'
 import { DeFiDModule } from '@src/module.defid'
 import { HealthModule } from '@src/module.health'
