+++ src/module.api/call.controller.ts
@@ -14,9 +14,9 @@ import {
 } from '@nestjs/common'
 
 import { JsonRpcClient } from '@defichain/jellyfish-api-jsonrpc'
-import { NetworkGuard } from '@src/module.api/commons/network.guard'
-import { ResponseInterceptor } from '@src/module.api/commons/response.interceptor'
-import { ExceptionInterceptor } from '@src/module.api/commons/exception.interceptor'
+import { NetworkGuard } from '@src/module.api/guards/network.guard'
+import { ResponseInterceptor } from '@src/module.api/interceptors/response.interceptor'
+import { ExceptionInterceptor } from '@src/module.api/interceptors/exception.interceptor'
 import { HttpExceptionFilter } from './commons/exception.filter'
 
 /**
