+++ src/module.api/interceptors/exception.interceptor.ts
